linux_boot: reserve space for the linux vectors
Writing to FLASH from linux requires placing some of the kernel into the
IRAM so that it could run while the FLASH cache is off. It looks like we
can't just use random part of IRAM when firmware is running on the other
core. Reserve two pages of IRAM to guarantee that there's a page-aligned
area for the kernel vectors, exception handlers and .xiptext.
Unfortunately the kernel must have the address of this area in its
configuration, so keep an eye on the map_flash_and_go symbol address.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>