linux_boot: align flash mapping with partition structure
FLASH partition mapping in the CPU address space does not match FLASH
partition mapping on the chip itself. This is awkward and unnecessarily
complex. Align FLASH partition mapping with the partition structure on
the FLASH chip. With this change partition with X in the 'Offset' column
of the partition table is mapped at offset X from the FLASH region base
address in the memory.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>